+++ packages/inline-mod/src/log.ts
@@ -1,10 +1,5 @@
 import debug from 'debug';
 
-debug.inspectOpts = {
-	colors: true,
-	depth: null,
-};
-
 const baseLogger = debug('inox-tools:im');
 
 export function getLogger(name: string): debug.Debugger {